\name{all.Amianhinfusionsm}

\alias{all.Amianhinfusionsm}

\title{Process to estimate individual PK parameters of aminophylline anhydrous iv infusion with multiple concentrations}

\description{
 Process to estimate individual PK parameters of aminophylline anhydrous iv infusion with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}