\encoding{UTF-8}

\name{all.Amianhirms}

\alias{all.Amianhirms}

\title{Process to estimate many subjects' PK parameters of aminophylline anhydrous IR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of aminophylline anhydrous IR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}