\name{all.Amianhirsm}

\alias{all.Amianhirsm}

\title{Process to estimate individual PK parameters of aminophylline anhydrous IR tablet with multiple concentrations}

\description{
 Process to estimate individual PK parameters of aminophylline anhydrous IR tablet with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}