\encoding{UTF-8}

\name{all.Amianhirss}

\alias{all.Amianhirss}

\title{Process to estimate individual PK parameters of aminophylline anhydrous IR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of aminophylline anhydrous IR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}