\name{all.Amicrss}

\alias{all.Amicrss}

\title{Process to estimate individual PK parameters of aminophylline anhydrous CR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of aminophylline anhydrous CR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}