\name{all.Amidihcrms}

\alias{all.Amidihcrms}

\title{Process to estimate many subjects' PK parameters of aminophylline dihydrous CR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of aminophylline dihydrous CR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}