\encoding{UTF-8}

\name{all.Amidihcrss}

\alias{all.Amidihcrss}

\title{Process to estimate individual PK parameters of aminophylline dihydrous CR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of aminophylline dihydrous CR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}