\encoding{UTF-8}

\name{all.Amidihirsm}

\alias{all.Amidihirsm}

\title{Process to estimate individual PK parameters of aminophylline dihydrous IR tablet with multiple concentrations}

\description{
 Process to estimate individual PK parameters of aminophylline dihydrous IR tablet with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}