\name{all.Amidihirss}

\alias{all.Amidihirss}

\title{Process to estimate individual PK parameters of aminophylline dihydrous IR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of aminophylline dihydrous IR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}