\name{all.Amimm}

\alias{all.Amimm}

\title{Process to estimate many subjects' PK parameters of aminoglycoside with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of aminoglycoside with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}