\name{all.Amism}

\alias{all.Amism}

\title{Process to estimate individual PK parameters of aminoglycoside with multiple concentrations}

\description{
 Process to estimate individual PK parameters of aminoglycoside with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}