\name{all.Amiss}

\alias{all.Amiss}

\title{Process to estimate individual PK parameters of aminoglycoside with one concentration}

\description{
 Process to estimate individual PK parameters of aminoglycoside with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}