\encoding{UTF-8}

\name{all.Carmm}

\alias{all.Carmm}

\title{Process to estimate many subjects' PK parameters of carbamazepine with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of carbamazepine with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}