\encoding{UTF-8}

\name{all.Carsm}

\alias{all.Carsm}

\title{Process to estimate individual PK parameters of carbamazepine with multiple concentrations}

\description{
 Process to estimate individual PK parameters of carbamazepine with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}