\encoding{UTF-8}

\name{all.ChiValmm}

\alias{all.ChiValmm}

\title{Process to estimate many subjects' PK parameters of valproate for children with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of valproate for children with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}