\name{all.ChiValms}

\alias{all.ChiValms}

\title{Process to estimate many subjects' PK parameters of valproate for children by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of valproate for children by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}