\name{all.ChiValsm}

\alias{all.ChiValsm}

\title{Process to estimate individual PK parameters of valproate for children with multiple concentrations}

\description{
 Process to estimate individual PK parameters of valproate for children with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}