\name{all.Digms}

\alias{all.Digms}

\title{Process to estimate many subjects' PK parameters of digoxin by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of digoxin by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}