\encoding{UTF-8}

\name{all.Digss}

\alias{all.Digss}

\title{Process to estimate individual PK parameters of digoxin with one concentration}

\description{
 Process to estimate individual PK parameters of digoxin with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}