\name{all.Evems}

\alias{all.Evems}

\title{Process to estimate many subjects' PK parameters of everolimus by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of everolimus by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}