\encoding{UTF-8}

\name{all.Imamm}

\alias{all.Imamm}

\title{Process to estimate many subjects' PK parameters of imatinib mesylate with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of imatinib mesylate with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}