\name{all.Imams}

\alias{all.Imams}

\title{Process to estimate many subjects' PK parameters of imatinib mesylate by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of imatinib mesylate by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}