\encoding{UTF-8}

\name{all.Imasm}

\alias{all.Imasm}

\title{Process to estimate individual PK parameters of imatinib mesylate with multiple concentrations}

\description{
 Process to estimate individual PK parameters of imatinib mesylate with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}