\name{all.Imass}

\alias{all.Imass}

\title{Process to estimate individual PK parameters of imatinib mesylate with one concentration}

\description{
 Process to estimate individual PK parameters of imatinib mesylate with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}