\name{all.Indmm}

\alias{all.Indmm}

\title{Process to estimate many subjects' PK parameters of indinavir with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of indinavir with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}