\name{all.Indms}

\alias{all.Indms}

\title{Process to estimate many subjects' PK parameters of indinavir by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of indinavir by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}