\encoding{UTF-8}

\name{all.Indsm}

\alias{all.Indsm}

\title{Process to estimate individual PK parameters of indinavir with multiple concentrations}

\description{
 Process to estimate individual PK parameters of indinavir with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}