\name{all.Indss}

\alias{all.Indss}

\title{Process to estimate individual PK parameters of indinavir with one concentration}

\description{
 Process to estimate individual PK parameters of indinavir with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}