\encoding{UTF-8}

\name{all.OO.sm}

\alias{all.OO.sm}

\title{Main function to run opioids tdm}

\description{
Main function to run opioids tdm
}
\references{
1. Linares, O and Linares, A. Computational Opioid Prescribing: A Novel Application of Clinical
   Pharmacokinetics. Journal of Pain and Palliative Care Pharmacotherapy. Vol 25, No 2, June 2011
2. Linares, O, Daly, D and Linares, A. Personalized Oxycodone Dosing: Using Pharmacogenetic 
   Testing and Clinical Pharmacokinetics to Reduce Toxicity Risk and Increase Effectiveness.
   Pain Medicine, 2014,...
}

\keyword{misc}
