\encoding{UTF-8}

\name{all.Oxtcrms}

\alias{all.Oxtcrms}

\title{Process to estimate many subjects' PK parameters of oxtriphylline CR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of oxtriphylline CR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}