\encoding{UTF-8}

\name{all.Oxtirmm}

\alias{all.Oxtirmm}

\title{Process to estimate many subjects' PK parameters of oxtriphylline IR tablet with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of oxtriphylline IR tablet with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}