\encoding{UTF-8}

\name{all.Oxtirms}

\alias{all.Oxtirms}

\title{Process to estimate many subjects' PK parameters of oxtriphylline IR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of oxtriphylline IR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}