\encoding{UTF-8}

\name{all.Oxtirss}

\alias{all.Oxtirss}

\title{Process to estimate individual PK parameters of oxtriphylline IR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of oxtriphylline IR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}