\encoding{UTF-8}

\name{all.Phems}

\alias{all.Phems}

\title{Process to estimate many subjects' PK parameters of phenytoin by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of phenytoin by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}