\name{all.Ritsm}

\alias{all.Ritsm}

\title{Process to estimate individual PK parameters of ritonavir with multiple concentrations}

\description{
 Process to estimate individual PK parameters of ritonavir with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}