\name{all.Ritss}

\alias{all.Ritss}

\title{Process to estimate individual PK parameters of ritonavir with one concentration}

\description{
 Process to estimate individual PK parameters of ritonavir with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}