\encoding{UTF-8}

\name{all.Tacsm}

\alias{all.Tacsm}

\title{Process to estimate individual PK parameters of tacrolimus with multiple concentrations}

\description{
 Process to estimate individual PK parameters of tacrolimus with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}