\name{all.Thecrms}

\alias{all.Thecrms}

\title{Process to estimate many subjects' PK parameters of theophylline CR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of theophylline CR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}