\name{all.Theirmm}

\alias{all.Theirmm}

\title{Process to estimate many subjects' PK parameters of theophylline IR tablet with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of theophylline IR tablet with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}