\encoding{UTF-8}

\name{all.Theirms}

\alias{all.Theirms}

\title{Process to estimate many subjects' PK parameters of theophylline IR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of theophylline IR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}