\name{all.Theirsm}

\alias{all.Theirsm}

\title{Process to estimate individual PK parameters of theophylline IR tablet with multiple concentrations}

\description{
 Process to estimate individual PK parameters of theophylline IR tablet with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}