\name{all.Theirss}

\alias{all.Theirss}

\title{Process to estimate individual PK parameters of theophylline IR tablet with one concentration}

\description{
 Process to estimate individual PK parameters of theophylline IR tablet with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}