\name{all.Vanmm}

\alias{all.Vanmm}

\title{Process to estimate many subjects' PK parameters of vancomycin with different concentrations}

\description{
 Process to estimate many subjects' PK parameters of vancomycin with different concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}