\encoding{UTF-8}

\name{all.Vanms}

\alias{all.Vanms}

\title{Process to estimate many subjects' PK parameters of vancomycin by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of vancomycin by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}