\encoding{UTF-8}

\name{all.Vansm}

\alias{all.Vansm}

\title{Process to estimate individual PK parameters of vancomycin with multiple concentrations}

\description{
 Process to estimate individual PK parameters of vancomycin with multiple concentrations at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}