\name{all.Vanss}

\alias{all.Vanss}

\title{Process to estimate individual PK parameters of vancomycin with one concentration}

\description{
 Process to estimate individual PK parameters of vancomycin with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}