\name{all.Warss}

\alias{all.Warss}

\title{Process to estimate individual PK parameters of warfarin with one concentration}

\description{
 Process to estimate individual PK parameters of warfarin with one concentration at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}