\encoding{UTF-8}

\name{Ami.model}

\alias{Ami.model}

\title{
 Algorithm to estimate individual PK parameters of aminoglycoside}

\description{
 Algorithm to estimate individual PK parameters of aminoglycoside by Bayesian 
 technique(OpenBUGS).
}

\references{
Burton ME, Chow MS, et al., (1986). Accuracy of Bayesian and Sawchuk-Zaske 
dosing methods for gentamicin. Clin Pharm 5(2): 143-9.
}

\keyword{misc}