\encoding{UTF-8}

\name{Ami.more}

\alias{Ami.more}

\title{Function for individual dosage adjustment of aminoglycoside}

\description{
 Function to predict individual concentrations of aminoglycoside or to 
 calculate their suggested doses after perdicting individual PK parameters
 with single concentration. 
}

\keyword{misc}