\name{Ami.ms}

\alias{Ami.ms}

\title{Algorithm to estimate many subjects' PK parameters of aminoglycoside by each one conc.}

\description{
 Algorithm to estimate many subjects' PK parameters of aminoglycoside by each one conc. at steady-state.
}

\references{
 Burton, M. E., M. S. Chow, et al. (1986). Accuracy of Bayesian and Sawchuk-Zaske dosing methods for 
 gentamicin. Clin Pharm 5(2): 143-9.
}

\keyword{misc}