\name{Ami.sm}

\alias{Ami.sm}

\title{Algorithm to estimate individual PK parameters of aminoglycoside with multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of aminoglycoside with multiple concentrations at steady-state.
}

\references{
 Burton, M. E., M. S. Chow, et al. (1986). Accuracy of Bayesian and Sawchuk-Zaske dosing methods for gentamicin. 
 Clin Pharm 5(2): 143-9. 
}

\keyword{misc}