\encoding{UTF-8}

\name{Ami.ss}

\alias{Ami.ss}

\title{Algorithm to estimate individual PK parameters of aminoglycoside with one concentration}

\description{
 Algorithm to estimate one subject's PK parameters of aminoglycoside with one sampling at steady-state.
}

\references{
Burton, M. E., M. S. Chow, et al. (1986). Accuracy of Bayesian and Sawchuk-Zaske dosing methods for 
gentamicin. Clin Pharm 5(2): 143-9. 
}

\keyword{misc}