\encoding{UTF-8}

\name{Amianhcr.model}

\alias{Amianhcr.model}

\title{Algorithm to estimate individual PK parameters of aminophylline anhydrous CR}

\description{
 Algorithm to estimate individual PK parameters of aminophylline anhydrous CR by Bayesian technique(JAGS).
}

\references{
 Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}