\name{Amianhcrms.more}

\alias{Amianhcrms.more}

\title{Function for multiple subjects' dosage adjustment of aminophylline dihydrous CR}

\description{
 Function to predict multiple subjects' concentrations of aminophylline dihydrous CR or to calculate their suggested doses.
}

\keyword{misc}