\name{all.Amianhcrms}

\alias{all.Amianhcrms}

\title{Process to estimate many subjects' PK parameters of aminophylline anhydrous CR tablet by each one conc.}

\description{
 Process to estimate many subjects' PK parameters of aminophylline anhydrous CR tablet by each one conc. at steady-state including estimation of PK parameters and dose adjustment.
}

\keyword{misc}