\name{Amianhcrmscd.output}

\alias{Amianhcrmscd.output}

\title{output file about predicting multiple subjects' PK parameters of aminophylline anhydrous CR and dose adjustment(C<->D)}

\description{
 output file about predicting multiple subjects' PK parameters of aminophylline anhydrous CR by single conc. 
 and dose adjustment(C<->D).
}

\keyword{misc}