\name{Amianhcrss.output}

\alias{Amianhcrss.output}

\title{output file about predicting individual PK parameters of aminophylline anhydrous CR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous CR by single conc. 
 and dose adjustment (C->D or D->C).
}

\keyword{misc}