\encoding{UTF-8}

\name{Amianhinfusion.model}

\alias{Amianhinfusion.model}

\title{Estimate individual PK parameters of iv infusion aminophylline anhydrous with loading dose }

\description{
 Estimate individual PK parameters of iv infusion aminophylline anhydrous 
 with loading dose by Bayesian technique(OpenBUGS).
}

\references{
Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the optimization 
of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}