\name{Amianhinfusion.more}

\alias{Amianhinfusion.more}

\title{Function for individual dosage adjustment of iv infusion aminophylline anhydrous}

\description{
 Function to predict individual concentrations of iv infusion aminophylline anhydrous or to calculate their suggested doses.
}

\keyword{misc}