\name{Amianhinfusion.sm}

\alias{Amianhinfusion.sm}

\title{Algorithm to estimate individual PK parameters of iv infusion aminophylline anhydrous 
with loading dose by multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of iv infusion aminophylline anhydrous 
 with loading dose with multiple concentrations at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization of theophylline 
 therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}