\name{Amianhinfusionms.more}

\alias{Amianhinfusionms.more}

\title{Function for multiple subjects' dosage adjustment of iv infusion aminophylline anhydrous}

\description{
 Function to predict multiple subjects' concentrations of iv infusion aminophylline anhydrous or to calculate their suggested doses.
}

\keyword{misc}