\name{Amianhir.mm}

\alias{Amianhir.mm}

\title{Algorithm to estimate many subjects' PK parameters of aminophylline anhydrous IR with different concentrations}

\description{
 Algorithm to estimate many subjects' PK parameters of aminophylline anhydrous IR with different 
 concentrations at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the
 optimization of theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}