\encoding{UTF-8}

\name{Amianhir.more}

\alias{Amianhir.more}

\title{Function for individual dosage adjustment of aminophylline anhydrous IR}

\description{
 Function to predict individual concentrations of aminophylline anhydrous IR or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.
}

\keyword{misc}