\name{Amianhirms.more}

\alias{Amianhirms.more}

\title{Function for multiple subjects' dosage adjustment of aminophylline anhydrous IR}

\description{
 Function to predict multiple subjects' concentrations of aminophylline anhydrous IR or to calculate their suggested doses.
}

\keyword{misc}