\name{Amianhirms.pkoutput}

\alias{Amianhirms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of aminophylline anhydrous IR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of aminophylline anhydrous IR by each single conc..
}

\keyword{misc}