\name{Amianhirsm.more}

\alias{Amianhirsm.more}

\title{Function for individual dosage adjustment of aminophylline anhydrous IR}

\description{
 Function to predict individual concentrations of aminophylline anhydrous IR or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}