\name{Amianhirsm.output}

\alias{Amianhirsm.output}

\title{output file about predicting individual PK parameters of aminophylline anhydrous IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous IR by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}