\name{Amianhirsm.pkoutput}

\alias{Amianhirsm.pkoutput}

\title{output file about predicting individual PK parameters of aminophylline anhydrous IR by multiple conc.}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous IR by multiple conc..
}

\keyword{misc}