\name{Amianhirsmcd.output}

\alias{Amianhirsmcd.output}

\title{output file about predicting individual PK parameters of aminophylline anhydrous IR and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous IR by multiple conc. and dose adjustment (C<->D).
}

\keyword{misc}