\name{Amianhirss.pkoutput}

\alias{Amianhirss.pkoutput}

\title{output file about predicting individual PK parameters of aminophylline anhydrous IR by single conc.}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous IR by single conc..
}

\keyword{misc}