\name{Amianhirsscd.output}

\alias{Amianhirsscd.output}

\title{output file about predicting individual PK parameters of aminophylline anhydrous IR and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of aminophylline anhydrous IR by single conc. and dose adjustment (C<->D).
}

\keyword{misc}