\encoding{UTF-8}

\name{Amidihcr.model}

\alias{Amidihcr.model}

\title{Algorithm to estimate individual PK parameters of aminophylline dihydrous CR}

\description{
 Algorithm to estimate individual PK parameters of aminophylline dihydrous CR by Bayesian technique(OpenBUGS).
}

\references{
 Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}