\name{Amidihcrms.pkoutput}

\alias{Amidihcrms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of aminophylline dihydrous CR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of aminophylline dihydrous CR by each single conc..
}

\keyword{misc}