\name{Amidihcrss.output}

\alias{Amidihcrss.output}

\title{output file about predicting individual PK parameters of aminophylline dihydrous CR and dose adjustment (C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous CR by single conc. 
 and dose adjustment (C->D or D->C).
}

\keyword{misc}