\name{Amidihcrss.pkoutput}

\alias{Amidihcrss.pkoutput}

\title{output file about predicting individual PK parameters of aminophylline dihydrous CR by single conc.}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous CR by single conc..
}

\keyword{misc}