\name{Amidihcrsscd.output}

\alias{Amidihcrsscd.output}

\title{output file about predicting individual PK parameters of aminophylline dihydrous CR and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous CR by single conc. and dose adjustment (C<->D).
}

\keyword{misc}