\name{Amidihinfusion.mm}

\alias{Amidihinfusion.mm}

\title{Algorithm to estimate many subjects' PK parameters of iv infusion aminophylline dihydrous 
with loading dose with each different concentrations}

\description{
 Algorithm to estimate many subjects' PK parameters of iv infusion aminophylline dihydrous 
 with loading dose with each different concentrations at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population  to the optimization of theophylline therapy. 
 J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}