\name{Amidihinfusion.more}

\alias{Amidihinfusion.more}

\title{Function for individual dosage adjustment of iv infusion aminophylline dihydrous}

\description{
 Function to predict individual concentrations of iv infusion Aminophylline dihydrous or to calculate their suggested doses. 
}


\keyword{misc}