\encoding{UTF-8}

\name{Amidihinfusion.ms}

\alias{Amidihinfusion.ms}

\title{Algorithm to estimate many subjects' PK parameters of iv infusion aminophylline dihydrous 
with loading dose by each one conc.}

\description{
 Algorithm to estimate many subjects' PK parameters of iv infusion aminophylline dihydrous 
 with loading dose by each one conc.  at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization of 
 theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}