\name{Amidihinfusion.ss}

\alias{Amidihinfusion.ss}

\title{Algorithm to estimate individual PK parameters of iv infusion aminophylline dihydrous}

\description{
 Algorithm to estimate individual PK parameters of iv infusion aminophylline dihydrous with one sampling at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization of 
 theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}