\name{Amidihinfusionms.more}

\alias{Amidihinfusionms.more}

\title{Function for multiple subjects' dosage adjustment of iv infusion aminophylline dihydrous}

\description{
 Function to predict multiple subjects' concentrations of iv infusion Aminophylline dihydrous or to calculate their suggested doses. 
}

\keyword{misc}