\name{Amidihinfusionsm.more}

\alias{Amidihinfusionsm.more}

\title{Function for individual dosage adjustment of iv infusion aminophylline dihydrous with loading dose}

\description{
 Function to predict individual concentrations of iv infusion aminophylline dihydrous with loading dose 
 or to calculate their suggested doses after perdicting individual PK parameters with multiple conc.. 
}

\keyword{misc}