\name{Amidihinfusionsm.output}

\alias{Amidihinfusionsm.output}

\title{output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
with loading dose and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
 with loading dose by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}