\name{Amidihinfusionsmcd.output}

\alias{Amidihinfusionsmcd.output}

\title{output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
with loading dose and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of iv infusion aminophylline dihydrous with 
 loading dose by multiple concentrations and dose adjustment (C<->D).
}

\keyword{misc}