\name{Amidihinfusionss.pkoutput}

\alias{Amidihinfusionss.pkoutput}

\title{output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
with loading dose by single conc.}

\description{
 output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
 with loading dose by single conc..
}

\keyword{misc}