\name{Amidihinfusionsscd.output}

\alias{Amidihinfusionsscd.output}

\title{output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
with loading dose and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of iv infusion aminophylline dihydrous 
 with loading dose by single conc. 
 and does adjustment(C<->D).
}

\keyword{misc}