\name{Amidihir.model}

\alias{Amidihir.model}

\title{Algorithm to estimate individual PK parameters of aminophylline dihydrous IR}

\description{
 Algorithm to estimate individual PK parameters of aminophylline dihydrous IR by Bayesian technique(OpenBUGS).
}

\references{
 Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}