\encoding{UTF-8}

\name{Amidihir.more}

\alias{Amidihir.more}

\title{Function for individual dosage adjustment of aminophylline dihydrous IR}

\description{
 Function to predict individual concentrations of aminophylline dihydrous IR or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}