\name{Amidihir.sm}

\alias{Amidihir.sm}

\title{Algorithm to estimate individual PK parameters of aminophylline dihydrous IR with multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of aminophylline dihydrous IR with multiple concentrations 
 at steady-state.
 } 

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization of 
 theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}