\name{Amidihirms.more}

\alias{Amidihirms.more}

\title{Function for multiple subjects' dosage adjustment of aminophylline dihydrous IR}

\description{
 Function to predict multiple subjects' concentrations of aminophylline dihydrous IR or to calculate their suggested doses.  
}

\keyword{misc}