\name{Amidihirms.pkoutput}

\alias{Amidihirms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of aminophylline dihydrous IR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of aminophylline dihydrous IR by each single conc..
}

\keyword{misc}