\name{Amidihirsm.more}

\alias{Amidihirsm.more}

\title{Function for individual dosage adjustment of aminophylline dihydrous IR}

\description{
 Function to predict individual concentrations of aminophylline dihydrous IR or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..  
}

\keyword{misc}