\name{Amidihirsm.output}

\alias{Amidihirsm.output}

\title{output file about predicting individual PK parameters of aminophylline dihydrous IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous IR by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}