\name{Amidihirsm.pkoutput}

\alias{Amidihirsm.pkoutput}

\title{output file about predicting individual PK parameters of aminophylline dihydrous IR by multiple conc.}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous IR by multiple conc..
}

\keyword{misc}