\name{Amidihirsmcd.output}

\alias{Amidihirsmcd.output}

\title{output file about predicting individual PK parameters of aminophylline dihydrous IR and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous IR by multiple conc. and dose adjustment(C<->D).
}

\keyword{misc}