\name{Amidihirss.output}

\alias{Amidihirss.output}

\title{output file about predicting individual PK parameters of aminophylline dihydrous IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous IR by single conc. 
 and dose adjustment (C->D or D->C).
}

\keyword{misc}