\name{Amidihirss.pkoutput}

\alias{Amidihirss.pkoutput}

\title{output file about predicting individual PK parameters of aminophylline dihydrous IR by single conc.}

\description{
 output file about predicting individual PK parameters of aminophylline dihydrous IR by single conc..
}

\keyword{misc}