\name{Amims.pkoutput}

\alias{Amims.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of aminoglycoside by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of aminoglycoside by each single conc..
}

\keyword{misc}