\encoding{UTF-8}

\name{Amism.more}

\alias{Amism.more}

\title{Function for individualdosage adjustment of aminoglycoside}

\description{
 Function to predict individual concentrations of aminoglycoside or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc.. 
}

\keyword{misc}