\name{Amismcd.output}

\alias{Amismcd.output}

\title{output file about predicting individual PK parameters of aminoglycoside and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of aminoglycoside by multiple conc. and dose adjustment(C<->D).
}

\keyword{misc}