\name{Amiss.output}

\alias{Amiss.output}

\title{output file about predicting individual PK parameters of aminoglycoside and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of aminoglycoside by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}