\name{Amisscd.output}

\alias{Amisscd.output}

\title{output file about predicting individual PK parameters of aminoglycoside and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of aminoglycoside by single conc. and dose adjustment(C<->D).
}

\keyword{misc}