\name{Car.mm}

\alias{Car.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of carbamazepine with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of carbamazepine with different 
 concentrations at steady-state.
}

\references{
 Jiao, Z., M. K. Zhong, et al. (2003). Population pharmacokinetics of carbamazepine in
 Chinese epilepsy patients. Ther Drug Monit 25(3): 279-86.
}

\keyword{misc}