\name{Car.model}

\alias{Car.model}

\title{Algorithm to estimate individual PK parameters of carbamazepine}

\description{
 Algorithm to estimate individual PK parameters of carbamazepine by Bayesian technique(OpenBUGS).
}

\references{
 Jiao, Z., M. K. Zhong, et al. (2003). Population pharmacokinetics of carbamazepine in Chinese 
 epilepsy patients. Ther Drug Monit 25(3): 279-86. 
}

\keyword{misc}