\encoding{UTF-8}

\name{Car.ms}

\alias{Car.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of carbamazepine by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of carbamazepine by each one conc. at steady-state.
}

\references{
 Jiao, Z., M. K. Zhong, et al. (2003). Population pharmacokinetics of carbamazepine in Chinese epilepsy 
 patients. Ther Drug Monit 25(3): 279-86.  
}

\keyword{misc}