\name{Car.sm}

\alias{Car.sm}

\title{Algorithm to estimate individual PK parameters of carbamazepine with multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of carbamazepine with multiple concentrations 
 at steady-state.
}

\references{
 Jiao, Z., M. K. Zhong, et al. (2003). Population pharmacokinetics of carbamazepine in Chinese 
 epilepsy patients. Ther Drug Monit 25(3): 279-86. 
}

\keyword{misc}