\name{Carsm.more}

\alias{Carsm.more}

\title{Function for individual dosage adjustment of carbamazepine}

\description{
 Function to predict individual concentrations of carbamazepine or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}