\name{Carsmcd.output}

\alias{Carsmcd.output}

\title{output file about predicting individual PK parameters of carbamazepine and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of carbamazepine by multiple conc. and dose adjustment(C<->D).
}

\keyword{misc}