\name{Carss.output}

\alias{Carss.output}

\title{output file about predicting individual PK parameters of carbamazepine and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of carbamazepine by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}