\name{Carsscd.output}

\alias{Carsscd.output}

\title{output file about predicting individual PK parameters of carbamazepine and dose adjustment (C<->D)}

\description{
 output file about predicting individual PK parameters of carbamazepine by single conc. and dose adjustment(C<->D).
}

\keyword{misc}