\name{ChiVal.mm}

\alias{ChiVal.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of valproate in children with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of valproate in children with differnet 
 concentrations at steady-state.
}

\references{
 B. Blanco Serrano et al, Valproate population pharmacokinetics in children. Journal of clinical
 pharmacy and Therapeutics (1999) 24,73-80.
}

\keyword{misc}