\encoding{UTF-8}

\name{ChiVal.model}

\alias{ChiVal.model}

\title{Algorithm to estimate individual PK parameters of valproate in children}

\description{
 Algorithm to estimate individual PK parameters of valproate in children by Bayesian technique(JAGS).
}

\references{
 B. Blanco Serrano et al, Valproate population pharmacokinetics in children. Journal of clinical pharmacy 
 and Therapeutics (1999) 24,73-80.
}

\keyword{misc}