\name{ChiVal.ms}

\alias{ChiVal.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of valproate in children by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of valproate in children by each one conc. at steady-state.
}

\references{
 B. Blanco Serrano et al, Valproate population pharmacokinetics in children. Journal of clinical pharmacy and 
 Therapeutics (1999) 24,73-80.
}

\keyword{misc}