\name{ChiVal.sm}

\alias{ChiVal.sm}

\title{Algorithm to estimate individual PK parameters of valproate in child with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of valproate in child with each multiple concentrations
 at steady-state.
}

\references{
 B. Blanco Serrano et al, Valproate population pharmacokinetics in children. Journal of clinical pharmacy 
 and Therapeutics (1999) 24,73-80.
}

\keyword{misc}