\name{ChiValsm.more}

\alias{ChiValsm.more}

\title{Function for individual dosage adjustment of valproic acid in children}

\description{
 Function to predict individual concentrations of valproic acid in children or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}