\name{ChiValsm.output}

\alias{ChiValsm.output}

\title{output file about predicting child PK parameters of valproate and dose adjustment(C->D or D->C)}

\description{
 output file about predicting child PK parameters of valproate by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}