\name{ChiValsscd.output}

\alias{ChiValsscd.output}

\title{output file about predicting child PK parameters of valproate and dose adjustment(C<->D)}

\description{
 output file about predicting child PK parameters of valproate by single conc. and dose adjustment (C<->D).
}

\keyword{misc}