\encoding{UTF-8}

\name{Cyc.model}

\alias{Cyc.model}

\title{Algorithm to estimate individual PK parameters of cyclosporine-A}

\description{
 Algorithm to estimate individual PK parameters of cyclosporine-A by Bayesian technique(OpenBUGS).
}

\references{
 Parke, J. and B. G. Charles (1998). NONMEM population pharmacokinetic modeling of orally administered 
 cyclosporine from routine drug monitoring data after heart transplantation. Ther Drug Monit 20(3): 
 284-93. 
}

\keyword{misc}