\encoding{UTF-8}

\name{Cyc.more}

\alias{Cyc.more}

\title{Function for individual dosage adjustment of cyclosporine}

\description{
Function to predict individual concentrations of cyclosporine or to calculate their suggested doses
after perdicting individual PK parameters with single concentration. 
}

\keyword{misc}