\name{Cyc.ms}

\alias{Cyc.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of cyclosporine-A by each one conc.
}

\description{
 Algorithm to estimate multiple subjects' PK parameters of cyclosporine-A by each one conc. at steady-state.
}

\references{
 Parke, J. and B. G. Charles (1998). NONMEM population pharmacokinetic modeling of orally administered 
 cyclosporine from routine drug monitoring data after heart transplantation. Ther Drug Monit 20(3): 284-93.
}

\keyword{misc}