\name{Cyc.sm}

\alias{Cyc.sm}

\title{Algorithm to estimate individual PK parameters of cyclosporine-A with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of cyclosporine-A with each multiple concentrations 
 at steady-state.
}

\references{
 Parke, J. and B. G. Charles (1998). NONMEM population pharmacokinetic modeling of orally administered 
 cyclosporine from routine drug monitoring data after heart transplantation. Ther Drug Monit 20(3): 284-93. 
}

\keyword{misc}