\name{Cycsm.output}

\alias{Cycsm.output}

\title{output file about predicting individual PK parameters of cyclosporine-A and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of cyclosporine-A by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}