\name{Cyctwo.more}

\alias{Cyctwo.more}

\title{Function for individual dosage adjustment of cyclosporine}

\description{
Function to predict individual concentrations of cyclosporine or to calculate their suggested doses
after perdicting individual PK parameters with multiple conc.. 
}

\keyword{misc}