\encoding{UTF-8}

\name{Dig.model}

\alias{Dig.model}

\title{Algorithm to estimate individual PK parameters of digoxin}

\description{
 Algorithm to estimate individual PK parameters of digoxin by Bayesian technique(JAGS).
}

\references{
 Yukawa, E., H. Mine, et al. (1992). Digoxin population pharmacokinetics from routine clinical 
 data: role of patient characteristics for estimating dosing regimens. J Pharm Pharmacol 
 44(9): 761-5.  
}

\keyword{misc}