\encoding{UTF-8}

\name{Dig.more}

\alias{Dig.more}

\title{Function for individual dosage adjustment of digoxin}

\description{
 Function to predict individual concentrations of digoxin or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.   
}

\keyword{misc}