\name{Dig.ms}

\alias{Dig.ms}

\title{Algorithm to estimate many subjects PK parameters of digoxin by each one conc.}

\description{
 Algorithm to estimate many subjects PK parameters of digoxin by each one conc. at steady-state.
}

\references{
 Yukawa, E., H. Mine, et al. (1992). Digoxin population pharmacokinetics from routine clinical 
 data: role of patient characteristics for estimating dosing regimens. J Pharm Pharmacol 44(9): 
 761-5.
}

\keyword{misc}