\encoding{UTF-8}

\name{Dig.ss}

\alias{Dig.ss}

\title{Algorithm to estimate individual PK parameters of digoxin with one concentration}

\description{
 Algorithm to estimate individual PK parameters of digoxin with one concentration at steady-state.
}

\references{
 Yukawa, E., H. Mine, et al. (1992). Digoxin population pharmacokinetics from routine clinical 
 data: role of patient characteristics for estimating dosing regimens. J Pharm Pharmacol 44(9): 
 761-5. 
}

\keyword{misc}