\name{Digss.output}

\alias{Digss.output}

\title{output file about predicting individual PK parameters of digoxin and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of digoxin by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}