\encoding{UTF-8}

\name{Enf.model}

\alias{Enf.model}

\title{Algorithm to estimate individual PK parameters of enfuvirtide}

\description{
 Algorithm to estimate individual PK parameters of enfuvirtide by Bayesian technique(OpenBUGS).
}

\references{
 Mould, D. R., X. Zhang, et al. (2005). Population pharmacokinetics and exposure-response relationship 
 of enfuvirtide in treatment-experienced human immunodeficiency virus type 1-infected patients. Clin 
 Pharmacol Ther 77(6): 515-28. 
}

\keyword{misc}