\encoding{UTF-8}

\name{Enf.ms}

\alias{Enf.ms}

\title{Algorithm to estimate many subjects' PK parameters of enfuvirtide by each one conc.}

\description{
 Algorithm to estimate many subjects' PK parameters of enfuvirtide by each one conc. at steady-state.
}

\references{
 Mould, D. R., X. Zhang, et al. (2005). Population pharmacokinetics and exposure-response relationship 
 of enfuvirtide in treatment-experienced human immunodeficiency virus type 1-infected patients. Clin 
 Pharmacol Ther 77(6): 515-28.
}

\keyword{misc}