\name{Enf.sm}

\alias{Enf.sm}

\title{Algorithm to estimate individual PK parameters of enfuvirtide with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of enfuvirtide with each multiple concentrations 
 at steady-state.
}

\references{
 Mould, D. R., X. Zhang, et al. (2005). Population pharmacokinetics and exposure-response relationship 
 of enfuvirtide in treatment-experienced human immunodeficiency virus type 1-infected patients. Clin 
 Pharmacol Ther 77(6): 515-28. 
}

\keyword{misc}