\name{Enfsm.more}

\alias{Enfsm.more}

\title{Function for individual dosage adjustment of enfuvirtide}

\description{
 Function to predict individual concentrations of enfuvirtide or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..   
}

\keyword{misc}