\name{Eno.model}

\alias{Eno.model}

\title{Algorithm to estimate individual PK parameters of enoxaparin}

\description{
 Algorithm to estimate individual PK parameters of enoxaparin by Bayesian technique(OpenBUGS).
}

\references{
 Hulot, J. S., G. Montalescot, et al. (2005). Dosing strategy in patients with renal failure 
 receiving enoxaparin for the treatment of non-ST-segment elevation acute coronary syndrome. 
 Clin Pharmacol Ther 77(6): 542-52. 
}

\keyword{misc}