\name{Eno.more}

\alias{Eno.more}

\title{Function for individual dosage adjustment of enoxaparin}

\description{
 Function to predict individual anti-Xa maximal activit of enoxaparin or to calculate their suggested doses
 after perdicting individual PK parameters with single anti-Xa maximal activit.  
}

\keyword{misc}