\name{Eno.ms}

\alias{Eno.ms}

\title{Algorithm to estimate many subjects' PK parameters of enoxaparin by each one conc.}

\description{
 Algorithm to estimate many subjects' PK parameters of enoxaparin by each one conc. at steady-state.
}

\references{
 Hulot, J. S., G. Montalescot, et al. (2005). Dosing strategy in patients with renal failure receiving 
 enoxaparin for the treatment of non-ST-segment elevation acute coronary syndrome. Clin Pharmacol Ther 
 77(6): 542-52.
}

\keyword{misc}