\encoding{UTF-8}

\name{Eno.ss}

\alias{Eno.ss}

\title{Algorithm to estimate individual PK parameters of enoxaparin with one concentration}

\description{
 Algorithm to estimate individual PK parameters of enoxaparin with one concentration at steady-state.
}

\references{
Hulot, J. S., G. Montalescot, et al. (2005). Dosing strategy in patients with renal failure receiving 
enoxaparin for the treatment of non-ST-segment elevation acute coronary syndrome. Clin Pharmacol Ther 
77(6): 542-52.
}

\keyword{misc}