\name{Enoss.output}

\alias{Enoss.output}

\title{output file about predicting individual PK parameters of enoxaparin and dose adjustment(Amax->D or D->Amax)}

\description{
 output file about predicting individual PK parameters of enoxaparin by single anti-Xa maximal activity(Amax) 
 and dose adjustment (Amax->D or D->Amax).
}

\keyword{misc}