\name{Eve.model}

\alias{Eve.model}

\title{Algorithm to estimate individual PK parameters of everolimus}

\description{
 Algorithm to estimate individual PK parameters of everolimus by Bayesian technique(OpenBUGS).
}

\references{
 Kovarik, J. M., C. H. Hsu, et al. (2001). Population pharmacokinetics of everolimus in de novo 
 renal transplant patients: impact of ethnicity and comedications. Clin Pharmacol Ther 70(3): 
 247-54. Cl=Cl/F Vd=Vd/F
}

\keyword{misc}