\encoding{UTF-8}

\name{Eve.more}

\alias{Eve.more}

\title{Function for individual dosage adjustment of everolimus}

\description{
 Function to predict individual concentrations of everolimus or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.   
}

\keyword{misc}