\encoding{UTF-8}

\name{Eve.ms}

\alias{Eve.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of everolimus by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of everolimus by each one conc. at steady-state.
}

\references{
Kovarik, J. M., C. H. Hsu, et al. (2001). Population pharmacokinetics of everolimus in de novo renal 
transplant patients: impact of ethnicity and comedications. Clin Pharmacol Ther 70(3): 247-54.
}

\keyword{misc}