\name{Eve.sm}

\alias{Eve.sm}

\title{Algorithm to estimate individual PK parameters of everolimus with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of everolimus with each multiple concentrations 
 at steady-state.
 }

\references{
 Kovarik, J. M., C. H. Hsu, et al. (2001). Population pharmacokinetics of everolimus in de novo 
 renal transplant patients: impact of ethnicity and comedications. Clin Pharmacol Ther 70(3): 247-54. 
}

\keyword{misc}