\name{Evesm.more}

\alias{Evesm.more}

\title{Function for individual dosage adjustment of everolimus}

\description{
 Function to predict individual concentrations of everolimus or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..   
}

\keyword{misc}