\encoding{UTF-8}

\name{Ima.mm}

\alias{Ima.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of imatinib mesylate with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of imatinib mesylate with different 
 concentrations and ts must >1.5hr(absorption time).
}

\references{
 Schmidli, H., B. Peng, et al. (2005). Population pharmacokinetics of imatinib mesylate 
 in patients with chronic-phase chronic myeloid leukaemia: results of a phase III study. 
 Br J Clin Pharmacol 60(1): 35-44
}

\keyword{misc}