\encoding{UTF-8}

\name{Ima.model}

\alias{Ima.model}

\title{Algorithm to estimate individual PK parameters of imatinib mesylate}

\description{
 Algorithm to estimate individual PK parameters of imatinib mesylate by Bayesian technique(OpenBUGS).
}

\references{
 Schmidli, H., B. Peng, et al. (2005). Population pharmacokinetics of imatinib mesylate in patients 
 with chronic-phase chronic myeloid leukaemia: results of a phase III study. Br J Clin Pharmacol 
 60(1): 35-44
}

\keyword{misc}