\name{Imasm.more}

\alias{Imasm.more}

\title{Function for individual dosage adjustment of imatinib mesylate}

\description{
 Function to predict individual concentrations of imatinib mesylate or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}