\name{Imasm.output}

\alias{Imasm.output}

\title{output file about predicting individual PK parameters of imatinib mesylate and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of imatinib mesylate  by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}