\name{Imasscd.output}

\alias{Imasscd.output}

\title{output file about predicting individual PK parameters of imatinib mesylate and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of imatinib mesylate by single conc. and dose adjustment(C<->D).
}

\keyword{misc}