\encoding{UTF-8}

\name{Ind.mm}

\alias{Ind.mm}

\title{Algorithm to estimate many subjects' PK parameters of indinavir with each different concentrations}

\description{
 Algorithm to estimate many subjects' PK parameters of indinavir with different 
 concentrations at steady-state.
}

\references{
 Csajka, C., C. Marzolini, et al. (2004). Population pharmacokinetics of indinavir 
 in patients infected with human immunodeficiency virus. Antimicrob Agents Chemother 
 48(9): 3226-32. 
}

\keyword{misc}