\name{Ind.model}

\alias{Ind.model}

\title{Algorithm to estimate individual PK parameters of indinavir}

\description{
 Algorithm to estimate individual PK parameters of indinavir by Bayesian technique(OpenBUGS).
}

\references{
 Csajka, C., C. Marzolini, et al. (2004). Population pharmacokinetics of indinavir in patients 
 infected with human immunodeficiency virus. Antimicrob Agents Chemother 48(9): 3226-32. 
}

\keyword{misc}