\name{Ind.ms}

\alias{Ind.ms}

\title{Algorithm to estimate many subjects' PK parameters of indinavir by each one conc.}

\description{
 Algorithm to estimate many subjects' PK parameters of indinavir by each one conc. at steady-state.
}

\references{
 Csajka, C., C. Marzolini, et al. (2004). Population pharmacokinetics of indinavir in patients infected 
 with human immunodeficiency virus. Antimicrob Agents Chemother 48(9): 3226-32. 
}

\keyword{misc}