\encoding{UTF-8}

\name{Ind.sm}

\alias{Ind.sm}

\title{Algorithm to estimate individual PK parameters of indinavir with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of indinavir with each multiple concentrations 
 at steady-state.
}

\references{
 Csajka, C., C. Marzolini, et al. (2004). Population pharmacokinetics of indinavir in patients 
 infected with human immunodeficiency virus. Antimicrob Agents Chemother 48(9): 3226-32. 
}

\keyword{misc}