\encoding{UTF-8}

\name{Ind.ss}

\alias{Ind.ss}

\title{Algorithm to estimate individual PK parameters of indinavir with one concentration}

\description{
 Algorithm to estimate individual PK parameters of indinavir with one concentration at steady-state.
}

\references{
 Csajka, C., C. Marzolini, et al. (2004). Population pharmacokinetics of indinavir in patients infected 
 with human immunodeficiency virus. Antimicrob Agents Chemother 48(9): 3226-32. 
}

\keyword{misc}