\name{Indsm.more}

\alias{Indsm.more}

\title{Function for individual dosage adjustment of indinavir}

\description{
 Function to predict individual concentrations of indinavir or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc.. 
}

\keyword{misc}