\name{Indsm.output}

\alias{Indsm.output}

\title{output file about predicting individual PK parameters of indinavir and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of indinavir by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}