\name{Indsscd.output}

\alias{Indsscd.output}

\title{output file about predicting individual PK parameters of indinavir and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of indinavir by single conc. and dose adjustment(C<->D).
}

\keyword{misc}