% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{inv_log}
\alias{inv_log}
\title{Inverse Log2 of a Single Value}
\usage{
inv_log(value)
}
\arguments{
\item{value}{-- A numeric value.}
}
\value{
the inverse log of value.
}
\description{
Takes the inverse log of a value.
}

