% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{inv_log_transform}
\alias{inv_log_transform}
\title{Inverse Log Transform}
\usage{
inv_log_transform(data = NULL, file = NULL)
}
\arguments{
\item{data}{-- A data.table of values. The first column should be
gene symbols and the rest should be expression values.}

\item{file}{-- A filename with expression values. This is an alternative
to specifying data.}
}
\value{
the transformed data.table.
}
\description{
Given a data.table of values, each value is inverse log transformed.
}

