% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{iqr_dt}
\alias{iqr_dt}
\title{IQR}
\usage{
iqr_dt(datatable)
}
\arguments{
\item{datatable}{-- A data.table whose first column will be ignored
but whose other columns are all numeric.}
}
\value{
IQR of the data.table.
}
\description{
Finds the IQR of a data.table.
}

