\encoding{UTF-8}

\name{Lit.more}

\alias{Lit.more}

\title{Function for individual dosage adjustment of lithium carbonate}

\description{
 Function to predict individual concentrations of lithium carbonate or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}