\encoding{UTF-8}

\name{Lit.ms}

\alias{Lit.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of lithium carbonate by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of lithium carbonate by each one conc. at steady-state.
}

\references{
 Yukawa, E., N. Nomiyama, et al. (1993). Lithium population pharmacokinetics from routine clinical data: 
 role of patient characteristics for estimating dosing regimens. Ther Drug Monit 15(2): 75-82.
}

\keyword{misc}