\encoding{UTF-8}

\name{Lit.ss}

\alias{Lit.ss}

\title{Algorithm to estimate individual PK parameters of lithium carbonate with one concentration}

\description{
 Algorithm to estimate individual PK parameters of lithium carbonate with one sampling at steady-state.
}

\references{
 Yukawa, E., N. Nomiyama, et al. (1993). Lithium population pharmacokinetics from routine clinical 
 data: role of patient characteristics for estimating dosing regimens. Ther Drug Monit 15(2): 75-82.
}

\keyword{misc}