\encoding{UTF-8}

\name{Litcit.model}

\alias{Litcit.model}

\title{Algorithm to estimate individual PK parameter of lithium citrate}

\description{
 Algorithm to estimate individual PK parameters of lithium citrate by Bayesian technique(OpenBUGS).
}

\references{
 Yukawa, E., N. Nomiyama, et al. (1993). Lithium population pharmacokinetics from routine clinical 
 data: role of patient characteristics for estimating dosing regimens. Ther Drug Monit 15(2): 75-82.
}

\keyword{misc}