\encoding{UTF-8}

\name{Litcit.more}

\alias{Litcit.more}

\title{Function for individual dosage adjustment of lithium citrate}

\description{
 Function to predict individual concentrations of lithium citrate or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration. 
}

\keyword{misc}