\name{Litcitss.output}

\alias{Litcitss.output}

\title{output file about predicting individual PK parameters of lithium citrate and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of lithium citrate by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}