\name{Litcitsscd.output}

\alias{Litcitsscd.output}

\title{output file about predicting individual PK parameters of lithium citrate and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of lithium citrate by single conc. and dose adjustment(C<->D).
}

\keyword{misc}