\name{Litss.output}

\alias{Litss.output}

\title{output file about predicting individual PK parameters of lithium carbonate and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of lithium carbonate by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}