\name{Litsscd.output}

\alias{Litsscd.output}

\title{output file about predicting PK parameters of lithium carbonate and dose adjustment (C<->D)}

\description{
 output file about predicting PK parameters of lithium carbonate with single subject and single conc. 
 and dose adjustment(C<->D).
}

\keyword{misc}