% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package_loader.R
\name{load_it}
\alias{load_it}
\title{load_it}
\usage{
load_it(pack, update = FALSE)
}
\arguments{
\item{pack}{-- A vector of character strings containing package names.}
}
\value{
nothing
}
\description{
Loads a character vector of packages, installing them from Bioconductor if necessary.
This will also load most CRAN packages.
}

