% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{log2p1}
\alias{log2p1}
\title{Log2 plus 1 a Single Value}
\usage{
log2p1(value)
}
\arguments{
\item{value}{-- A numeric value.}
}
\value{
the log2(1 + value)
}
\description{
Takes the log2 of 1 plus a value.
}

