% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{log_transform_p1}
\alias{log_transform_p1}
\title{Log2 Transform 1 Plus}
\usage{
log_transform_p1(data = NULL, file = NULL)
}
\arguments{
\item{data}{-- A data.table of values. The first column should be
gene symbols and the rest should be expression values.}

\item{file}{-- A filename with expression values. This is an alternative
to specifying data.}
}
\value{
the transformed data.table.
}
\description{
The input is log transformed such that all values are non-negative by
adding 1 to each value before transformation.
}

