% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{max_dt}
\alias{max_dt}
\title{Data.table MAX}
\usage{
max_dt(datatable)
}
\arguments{
\item{datatable}{-- A data.table whose first column will be ignored
but whose other columns are all numeric.}
}
\value{
the maximum numeric value in the data.table.
}
\description{
Finds the max of a data.table.
}

