% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{min_dt}
\alias{min_dt}
\title{Data.table MIN}
\usage{
min_dt(datatable)
}
\arguments{
\item{datatable}{-- A data.table whose first column will be ignored
but whose other columns are all numeric.}
}
\value{
the minimum numeric value in the data.table.
}
\description{
Finds the min of a data.table.
}

