\encoding{UTF-8}

\name{note_for_c_to_d_eno}

\alias{note_for_c_to_d_eno}

\title{Note for dose adjustment (C->D) for enoxaparin}

\description{
 Note for dose adjustment especially for enoxaparin includes anti-Xa maximal activity at stedy state (IU/mL), dose(IU) and dosing interval(hr).
}

\keyword{misc}