\encoding{UTF-8}

\name{note_for_ChiValss_output}

\alias{note_for_ChiValss_output}

\title{Note for child valproate output information}

\description{
 Output information of child valproate including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}