\encoding{UTF-8}

\name{note_for_d_to_c_eno}

\alias{note_for_d_to_c_eno}

\title{Note for dose adjustment (D->C) for enoxaparin}

\description{
 Note for dose adjustment especially for enoxaparin includes anti-Xa maximal activity at stedy state (IU/mL), dose(IU) and dosing interval(hr).
}

\keyword{misc}