\encoding{UTF-8}

\name{note_for_Dig_output}

\alias{note_for_Dig_output}

\title{Note for digoxin output information}

\description{
 Output information of digoxin including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}