\encoding{UTF-8}

\name{note_for_Enfss_output}

\alias{note_for_Enfss_output}

\title{Note for enfuvirtide output information}

\description{
 Output information of enfuvirtide including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}