\encoding{UTF-8}

\name{note_for_Imass_output}

\alias{note_for_Imass_output}

\title{Note for imatinib mesylate output information}

\description{
 Output information of imatinib mesylate including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}