\encoding{UTF-8}

\name{note_for_Indss_output}

\alias{note_for_Indss_output}

\title{Note for indinavir output information}

\description{
 Output information of indinavir including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}