\name{note_for_Lit_output}

\alias{note_for_Lit_output}

\title{Note for lithium carbonate output information}

\description{
 Output information of lithium carbonate including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}