\name{note_for_Litcit_output}

\alias{note_for_Litcit_output}

\title{Note for lithium citrate output information}

\description{
 Output information of lithium citrate including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}