\encoding{UTF-8}

\name{note_for_Ritss_output}

\alias{note_for_Ritss_output}

\title{Note for ritonavir output information}

\description{
 Output information of ritonavir including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}