\encoding{UTF-8}

\name{note_for_Tacss_output}

\alias{note_for_Tacss_output}

\title{Note for tacrolimus output information}

\description{
 Output information of tacrolimus including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}