\name{note_for_Thecr_output}

\alias{note_for_Thecr_output}

\title{Note for theophylline salt form CR tablet output information}

\description{
 Output information of theophylline salt form CR tablet including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}