\encoding{UTF-8}

\name{note_for_Theirsm_output}

\alias{note_for_Theirsm_output}

\title{Note for theophylline salt form IR tablet output information}
  
\description{
 Output information of theophylline salt form IR tablet including estimated PK parameters and predicted steady-state conc. 
}

\keyword{misc}