\encoding{UTF-8}

\name{note_for_Theirss_output}

\alias{note_for_Theirss_output}

\title{Note for theophylline salt form IR tablet output information}

\description{
 Output information of theophylline salt form IR tablet including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}