\encoding{UTF-8}

\name{note_for_Vanss_output}

\alias{note_for_Vanss_output}

\title{Note for vancomycin output information}

\description{
 Output information of vancomycin including estimated PK parameters, predicted steady-state conc. and predicted steady state measured conc. 
}

\keyword{misc}