\encoding{UTF-8}

\name{ocIR.more}

\alias{ocIR.more}

\title{Function for individual dosage adjustment of opiods tdm}

\description{
 Function to predict individual concentrations of opiods tdm or to calculate their suggested doses
 after perdicting individual PK parameters.  
}

\keyword{misc}