\encoding{UTF-8}

\name{OOsdcpr}

\alias{OOsdcpr}

\title{Function to calculate predicted conc. for opioids tdm}

\description{
Function to calculate predicted conc. for opioids tdm after the 1st dose
based on the estimated PK parameters; PK model is a one-compartment, with
1st-ordered absorption and the 1st-ordered elimination rate constants.
}

\keyword{misc}
