\encoding{UTF-8}

\name{OOsscpr}

\alias{OOsscpr}

\title{Function to calculate predicted conc. for opioids tdm
 at the steady-state}

\description{
Function to calculate predicted conc. at steady-state for opioids 
tdm based on the estimated PK parameters; PK model is a one-compartment, 
with 1st-ordered absorption and the 1st-ordered elimination rate 
constants.
}

\keyword{misc}
