\encoding{UTF-8}

\name{Oxtcr.model}

\alias{Oxtcr.model}

\title{Algorithm to estimate individual PK parameters of oxtriphylline CR}

\description{
 Algorithm to estimate individual PK parameters oxtriphyllline CR by Bayesian technique(JAGS).
}

\references{
Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the 
optimization of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}