\encoding{UTF-8}

\name{Oxtcr.more}

\alias{Oxtcr.more}

\title{Function for individual dosage adjustment of oxtriphylline CR}

\description{
 Function to predict individual concentrations of oxtriphylline CR or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}
\keyword{misc}