\name{Oxtcrms.pkoutput}

\alias{Oxtcrms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of oxtriphylline CR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of oxtriphylline CR by each single conc..
}

\keyword{misc}