\name{Oxtcrss.output}

\alias{Oxtcrss.output}

\title{output file about predicting individual PK parameters of oxtriphylline CR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of oxtriphylline CR by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}