\name{Oxtcrsscd.output}

\alias{Oxtcrsscd.output}

\title{output file about predicting individual PK parameters of oxtriphylline CR and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of oxtriphylline CR by single conc. and dose adjustment(C<->D).
}

\keyword{misc}