\encoding{UTF-8}

\name{Oxtir.mm}

\alias{Oxtir.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of oxtriphylline IR with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of oxtriphylline IR with different 
 concentrations at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to 
 the optimization of theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}