\encoding{UTF-8}

\name{Oxtir.model}

\alias{Oxtir.model}

\title{Algorithm to estimate individual PK parameters of oxtriphylline IR}

\description{
 Algorithm to estimate individual PK parameters of oxtriphylline IR by Bayesian technique(OpenBUGS).
}

\references{
 Otero MJ, Buelga DS, Vazquez MA, et al. Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther. Apr 1996;21(2):113-125.
}

\keyword{misc}