\encoding{UTF-8}

\name{Oxtir.more}

\alias{Oxtir.more}

\title{Function for individual dosage adjustment of oxtriphylline IR}

\description{
 Function to predict individual concentrations of oxtriphylline IR to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}