\name{Oxtirms.pkoutput}

\alias{Oxtirms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of oxtriphylline IR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of oxtriphylline IR by each single conc..
}

\keyword{misc}