\name{Oxtirsm.more}

\alias{Oxtirsm.more}

\title{Function for individual dosage adjustment of oxtriphylline IR}

\description{
 Function to predict individual concentrations of oxtriphylline IR to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..  
}

\keyword{misc}