\name{Oxtirsm.output}

\alias{Oxtirsm.output}

\title{output file about predicting individual PK parameters of oxtriphylline IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of oxtriphylline IR by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}