\name{Oxtirsmcd.output}

\alias{Oxtirsmcd.output}

\title{output file about predicting individual PK parameters of oxtriphylline IR and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of oxtriphylline IR by multiple conc. and dose adjustment(C<->D).
}

\keyword{misc}