\name{Oxtirss.output}

\alias{Oxtirss.output}

\title{output file about predicting individual PK parameters of oxtriphylline IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of oxtriphylline IR by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}