\name{Oxtirsscd.output}

\alias{Oxtirsscd.output}

\title{output file about predicting individual PK parameters of oxtriphylline IR and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of oxtriphylline IR by single conc. and dose adjustment(C<->D).
}

\keyword{misc}