\name{PedDig.model}

\alias{PedDig.model}

\title{Algorithm to estimate individual PK parameters of digoxin in pediatrics}

\description{
 Algorithm to estimate individual PK parameters of digoxin by Bayesian technique(OpenBUGS).
}

\references{
 Martin-Suarez, A., A. C. Falcao, et al. (2002). Population pharmacokinetics of digoxin in 
 pediatric patients. Ther Drug Monit 24(6): 742-5.
}

\keyword{misc}