\name{PedDig.more}

\alias{PedDig.more}

\title{Function for individual dosage adjustment of digoxin in pediatrics}

\description{
 Function to predict individual concentrations of digoxin in pediatrics or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}