\name{PedDig.ms}

\alias{PedDig.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of digoxin in pediatrics by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of digoxin in pediatrics by each one conc. at steady-state.
}

\references{
 Martin-Suarez, A., A. C. Falcao, et al. (2002). Population pharmacokinetics of digoxin in pediatric patients. 
 Ther Drug Monit 24(6): 742-5. 
}

\keyword{misc}