\name{PedDig.ss}

\alias{PedDig.ss}

\title{Algorithm to estimate individual PK parameters of digoxin in periatrics with one concentration}

\description{
 Algorithm to estimate individual PK parameters of digoxin in periatrics with one concentration at steady-state.
}

\references{
Martin-Suarez, A., A. C. Falcao, et al. (2002). Population pharmacokinetics of digoxin in pediatric patients. 
Ther Drug Monit 24(6): 742-5. 
}

\keyword{misc}