\name{PedDigss.output}

\alias{PedDigss.output}

\title{output file about predicting individual PK parameters of pediatric digoxin and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of pediatric digoxin by single conc. and dose adjustment(C->D or D->C).
}

\keyword{misc}