\name{PedDigsscd.output}

\alias{PedDigsscd.output}

\title{output file about predicting pediatric's PK parameters of digoxin and dose adjustment (C<->D)}

\description{
 output file about predicting pediatric's PK parameters of digoxin by single conc. and dose adjustment(C<->D).
}

\keyword{misc}