\name{Phe.model}

\alias{Phe.model}

\title{Algorithm to estimate individual PK parameters of phenytoin}

\description{
 Algorithm to estimate individual PK parameters of phenytoin by Bayesian technique(OpenBUGS).
}

\references{
Vozeh, S., K. T. Muir, et al. (1981). Predicting individual phenytoin dosage. J Pharmacokinet Biopharm 9(2): 131-46.
}

\keyword{misc}