\encoding{UTF-8}

\name{Phe.more}

\alias{Phe.more}

\title{Function for individual dosage adjustment of phenytoin}

\description{
 Function to predict individual concentrations of phenytoin or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.    
}

\keyword{misc}