\encoding{UTF-8}

\name{Phe.ms}

\alias{Phe.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of phenytoin by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of phenytoin by each one conc. at steady-state.
}

\references{
 Vozeh, S., K. T. Muir, et al. (1981). Predicting individual phenytoin dosage. J Pharmacokinet Biopharm 
 9(2): 131-46.
}

\keyword{misc}