\encoding{UTF-8}

\name{Phe.ss}

\alias{Phe.ss}

\title{Algorithm to estimate individual PK parameters of phenytoin with one concentration}

\description{
 Algorithm to estimate individual PK parameters of phenytoin with one concentration at steady-state.
}

\references{
Vozeh, S., Muir, K.T., et al. (1981). Predicting individual phenytoin dosage. J Pharmacokinet Biopharm 9(2): 131-46.
}

\keyword{misc}