\encoding{UTF-8}

\name{PheDemo}

\alias{PheDemo}

\title{Test runs to estimate individual PK parameters of phenytoin with one concentration}

\description{Test runs to estimate individual PK parameters of phenytoin with one concentration at steady-state including estimation of PK parameters and dose adjustment.}

\keyword{misc}