% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{quartiles_dt}
\alias{quartiles_dt}
\title{Quartiles}
\usage{
quartiles_dt(datatable)
}
\arguments{
\item{datatable}{-- A data.table whose first column will be ignored
but whose other columns are all numeric.}
}
\value{
1st and 3rd quartiles of the data.table.
}
\description{
Finds the 1st and 3rd quartiles of a data.table.
}

