\encoding{UTF-8}

\name{Rit.model}

\alias{Rit.model}

\title{Algorithm to estimate individual PK parameters of ritonavir}

\description{
 Algorithm to estimate individual PK parameters of ritonavir by Bayesian technique(OpenBUGS).
}

\references{
 Kappelhoff, B. S., A. D. Huitema, et al. (2005). Development and validation of a population 
 pharmacokinetic model for ritonavir used as a booster or as an antiviral agent in HIV-1-infected 
 patients. Br J Clin Pharmacol 59(2): 174-82.
}

\keyword{misc}