\encoding{UTF-8}

\name{Rit.sm}

\alias{Rit.sm}

\title{Algorithm to estimate individual PK parameters of ritonavir with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of ritonavir with each multiple concentrations 
 at steady-state.
}

\references{
 Kappelhoff, B. S., A. D. Huitema, et al. (2005). Development and validation of a population 
 pharmacokinetic model for ritonavir used as a booster or as an antiviral agent in HIV-1-infected 
 patients. Br J Clin Pharmacol 59(2): 174-82. 
}

\keyword{misc}