\name{Ritsm.more}

\alias{Ritsm.more}

\title{Function for individual dosage adjustment of ritonavire}

\description{
 Function to predict individual concentrations of ritonavir or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}