\name{Ritsm.output}

\alias{Ritsm.output}

\title{output file about predicting individual PK parameters of ritonavir and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of ritonavir by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}