\name{Ritsmcd.output}

\alias{Ritsmcd.output}

\title{output file about predicting individual PK parameters of ritonavir and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of ritonavir by multiple conc. and dose adjustment (C<->D).
}

\keyword{misc}