\name{Ritss.output}

\alias{Ritss.output}

\title{output file about predicting individual PK parameters of ritonavir and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of ritonavir by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}