\encoding{UTF-8}

\name{Tac.model}

\alias{Tac.model}

\title{Algorithm to estimate individual PK parameters of tacrolimus}

\description{
 Algorithm to estimate individual PK parameters of tacrolimus by Bayesian technique(OpenBUGS).
}

\references{
 Zahir, H., A. J. McLachlan, et al. (2005). Population pharmacokinetic estimation of tacrolimus 
 apparent clearance in adult liver transplant recipients. Ther Drug Monit 27(4): 422-30. 
}

\keyword{misc}