\name{Tac.more}

\alias{Tac.more}

\title{Function for individual dosage adjustment of tacrolimus}

\description{
 Function to predict individual concentrations of tacrolimus or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}