\encoding{UTF-8}

\name{Tac.ms}

\alias{Tac.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of tacrolimus by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of tacrolimus by each one conc. at steady-state.
}

\references{
 Zahir, H., A. J. McLachlan, et al. (2005). Population pharmacokinetic estimation of tacrolimus apparent 
 clearance in adult liver transplant recipients. Ther Drug Monit 27(4): 422-30.
}

\keyword{misc}