\encoding{UTF-8}

\name{Tac.sm}

\alias{Tac.sm}

\title{Algorithm to estimate individual PK parameters of tacrolimus with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of tacrolimus with each multiple concentrations 
 at steady-state.
}

\references{
 Zahir, H., A. J. McLachlan, et al. (2005). Population pharmacokinetic estimation of tacrolimus 
 apparent clearance in adult liver transplant recipients. Ther Drug Monit 27(4): 422-30. 
}

\keyword{misc}