\name{Tac.ss}

\alias{Tac.ss}

\title{Algorithm to estimate individual PK parameters of tacrolimus with one concentration}

\description{
 Algorithm to estimate individual PK parameters of tacrolimus with one concentration at steady-state.
}

\references{
 Zahir, H., A. J. McLachlan, et al. (2005). Population pharmacokinetic estimation of tacrolimus apparent 
 clearance in adult liver transplant recipients. Ther Drug Monit 27(4): 422-30. 
}

\keyword{misc}