\name{Tacsm.more}

\alias{Tacsm.more}

\title{Function for individual dosage adjustment of tacrolimus}

\description{
 Function to predict individual concentrations of tacrolimus or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..  
}

\keyword{misc}