\name{tdm}

\alias{tdm}

\title{Function of Therapeutic drug monitoring}

\description{
tdm is desgined to estimate individual PK parameters with one or more concentrations using
OpenBUGS (Bayesian inference Using Gibbs Sampling) interfaced through the package - BRugs.  
It also can calculate a suggested dose for a target drug conc. and vice versa.
}

\keyword{misc}
