% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{tdm_transform}
\alias{tdm_transform}
\title{TDM Transformation}
\usage{
tdm_transform(target_data = NULL, ref_data = NULL, file = NULL,
  ref_file = NULL, negative = FALSE, filter_p = FALSE,
  inv_reference = TRUE, log_target = TRUE, tdm_factor = 1)
}
\arguments{
\item{file}{-- The file of gene expression values to transform, presumably
containing RNA-seq values. It should have a header, but the header
should only contain column names, not a row name. All other rows should
have the same number of values and the first value should be a row name.
This is an inconveniance of working with data.table, but they are much faster.
Values should be tab-separated.}

\item{ref_file}{-- the file to transform in relation to, presumably microarray
expression values. Values should be tab-separated.}

\item{negative}{-- should the reference file be first inverse log transformed, then
log transformed again adding 1 to each value? Do this if the file has already
been log transformed and their are negative numbers in it.}

\item{filter_p}{-- should the patients in the target file be filtered to include
only those that match patients (column ids) in the reference file? Only the first
15 characters are checked for match.}

\item{inv_reference}{-- should reference data be inverse log transformed? Do this
if the target is not already log transformed (which is the typical case). This just
ensures that the comparison is meaningful.}

\item{log_target}{-- should the target file be log transformed after TDM? Usually
it should, since microarray data are typically log transformed and the whole
point of the transformation is to make the data more comparable.}

\item{tdm_factor}{-- a factor of the IQR, to adjust the granularity of distribution
matching.}
}
\value{
void
}
\description{
The expression files should have the same sorts of gene symbols. The target file
will be filtered to include only symbols that are in the reference file and the
order will also be changed to match. If a gene is not present in the target file,
an entry of all 0s will be added for that gene.
}

