\name{Thecr.more}

\alias{Thecr.more}

\title{Function for individual dosage adjustment of theophylline CR}

\description{
 Function to predict individual concentrations of theophylline CR or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration. 
}

\keyword{misc}