\encoding{UTF-8}

\name{Thecr.ms}

\alias{Thecr.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of theophylline CR by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of theophylline CR by each one conc. at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}