\name{Theinfusion.ss}

\alias{Theinfusion.ss}

\title{Algorithm to estimate individual PK parameters of iv infusion theophylline with one concentration}

\description{
 Algorithm to estimate individual PK parameters of iv infusion theophylline with one concentration at steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization of 
 theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}