\encoding{UTF-8}

\name{Theinfusioncpr}

\alias{Theinfusioncpr}

\title{Equation for the calculated conc. of iv infusion theophylline with loading dose at steady-state}

\description{
 Equation for the calculated conc. of iv infusion theophylline with loading dose at steady-state
}

\keyword{misc}