\name{Theinfusionms.more}

\alias{Theinfusionms.more}

\title{Function for multiple subjects' dosage adjustment of iv infusion theophylline}

\description{
 Function to predict multiple subjects' concentrations of  iv infusion theophylline or to calculate their suggested doses.  
}

\keyword{misc}