\name{Their.more}

\alias{Their.more}

\title{Function for individual dosage adjustment of theophylline IR}

\description{
 Function to predict individual concentrations of theophylline IR or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.
}

\keyword{misc}