\encoding{UTF-8}

\name{Their.sm}

\alias{Their.sm}

\title{Algorithm to estimate individual PK parameters of theophylline IR with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of theophylline IR with each multiple concentrations at 
 steady-state.
}

\references{
 Otero, M. J., D. S. Buelga, et al. (1996). Application of population pharmacokinetics to the optimization 
 of theophylline therapy. J Clin Pharm Ther 21(2): 113-25.
}

\keyword{misc}