\name{Theirms.pkoutput}

\alias{Theirms.pkoutput}

\title{output file about predicting multiple subjects' PK parameters of theophylline IR by each single conc.}

\description{
 output file about predicting multiple subjects' PK parameters of theophylline IR by each single conc..
}

\keyword{misc}