\name{Theirsm.output}

\alias{Theirsm.output}

\title{output file about predicting individual PK parameters of theophylline IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of theophylline IR by multiple conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}