\name{Theirsmcd.output}

\alias{Theirsmcd.output}

\title{output file about predicting individual PK parameters of theophylline IR and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of theophylline IR by multiple conc. and dose adjustment(C<->D).
}

\keyword{misc}