\name{Theirss.output}

\alias{Theirss.output}

\title{output file about predicting individual PK parameters of theophylline IR and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of theophylline IR by single conc. and dose adjustment (C->D or D->C).
}

\keyword{misc}