\name{Theirsscd.output}

\alias{Theirsscd.output}

\title{output file about predicting individual PK parameters of theophylline IR and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of theophylline IR by single conc. and dose adjustment(C<->D).
}

\keyword{misc}