\name{Val.mm}

\alias{Val.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of valproate with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of valproate with different 
 concentrations at steady-state.
}

\references{
 ES, E. L. D., E. Fuseau, et al. (2004). Pharmacokinetic modelling of valproic acid 
 from routine clinical data in Egyptian epileptic patients. Eur J Clin Pharmacol 
 59(11): 783-90. 
}

\keyword{misc}