\name{Val.model}

\alias{Val.model}

\title{Algorithm to estimate individual PK parameters of valproate}

\description{
 Algorithm to estimate individual PK parameters of valproate by Bayesian technique(OpenBUGS).
}

\references{
 ES, E. L. D., E. Fuseau, et al. (2004). Pharmacokinetic modelling of valproic acid from routine 
 clinical data in Egyptian epileptic patients. Eur J Clin Pharmacol 59(11): 783-90. 
}

\keyword{misc}