\encoding{UTF-8}

\name{Val.ms}

\alias{Val.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of valproate by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of valproate by each one conc. at steady-state.
}

\references{
 ES, E. L. D., E. Fuseau, et al. (2004). Pharmacokinetic modelling of valproic acid from routine clinical 
 data in Egyptian epileptic patients. Eur J Clin Pharmacol 59(11): 783-90.
}

\keyword{misc}