\encoding{UTF-8}

\name{Val.ss}

\alias{Val.ss}

\title{Algorithm to estimate individual PK parameters of valproate with one concentration}

\description{
 Algorithm to estimate individual PK parameters of valproate with one concentration at steady-state.
}

\references{
 ES, E. L. D., E. Fuseau, et al. (2004). Pharmacokinetic modelling of valproic acid from routine 
 clinical data in Egyptian epileptic patients. Eur J Clin Pharmacol 59(11): 783-90. 
}

\keyword{misc}