\name{Valtwo.more}

\alias{Valtwo.more}

\title{Function for individual dosage adjustment of valproic acid}

\description{
 Function to predict individual concentrations of vaproic acid or to calculate their suggested doses
 after perdicting individual PK parameters with multiple conc..
}

\keyword{misc}