\name{Van.mm}

\alias{Van.mm}

\title{Algorithm to estimate multiple subjects' PK parameters of vancomycin with different concentrations}

\description{
 Algorithm to estimate multiple subjects' PK parameters of vancomycin with different 
 concentrations at steady-state.
}

\references{
 Buelga, D. S., M. del Mar Fernandez de Gatta, et al. (2005). Population pharmacokinetic 
 analysis of vancomycin in patients with hematological malignancies. Antimicrob Agents 
 Chemother 49(12): 4934-41.
}

\keyword{misc}