\name{Van.more}

\alias{Van.more}

\title{Function for individual dosage adjustment of vancomycin}

\description{
 Function to predict individual concentrations of vancomycin or to calculate their suggested doses
 after perdicting individual PK parameters with single concentration.  
}

\keyword{misc}