\encoding{UTF-8}

\name{Van.ms}

\alias{Van.ms}

\title{Algorithm to estimate multiple subjects' PK parameters of vancomycin by each one conc.}

\description{
 Algorithm to estimate multiple subjects' PK parameters of vancomycin by each one conc. at steady-state.
}

\references{
 Buelga, D. S., M. del Mar Fernandez de Gatta, et al. (2005). Population pharmacokinetic analysis of 
 vancomycin in patients with hematological malignancies. Antimicrob Agents Chemother 49(12): 4934-41. 
}

\keyword{misc}