\name{Van.sm}

\alias{Van.sm}

\title{Algorithm to estimate individual PK parameters of vancomycin with each multiple concentrations}

\description{
 Algorithm to estimate individual PK parameters of vancomycin with each multiple concentrations at 
 steady-state.
}

\references{
 Buelga, D. S., M. del Mar Fernandez de Gatta, et al. (2005). Population pharmacokinetic analysis of 
 vancomycin in patients with hematological malignancies. Antimicrob Agents Chemother 49(12): 4934-41. 
}

\keyword{misc}