\encoding{UTF-8}

\name{Van.ss}

\alias{Van.ss}

\title{Algorithm to estimate individual PK parameters of vancomycin with one concentration}

\description{
 Algorithm to estimate individual PK parameters of vancomycin with one concentration at steady-state.
}

\references{
 Buelga, D. S., M. del Mar Fernandez de Gatta, et al. (2005). Population pharmacokinetic analysis of 
 vancomycin in patients with hematological malignancies. Antimicrob Agents Chemother 49(12): 4934-41. 
}

\keyword{misc}