\name{Vansm.output}

\alias{Vansm.output}

\title{output file about predicting individual PK parameters of vancomycin and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of vancomycin by multiple conc. and dose adjustment(C->D or D->C).
}

\keyword{misc}