\name{Vanss.output}

\alias{Vanss.output}

\title{output file about predicting individual PK parameters of vancomycin and dose adjustment(C->D or D->C)}

\description{
 output file about predicting individual PK parameters of vancomycin by single conc. and dose adjustment(C->D or D->C).
}

\keyword{misc}