\name{Vansscd.output}

\alias{Vansscd.output}

\title{output file about predicting individual PK parameters of vancomycin and dose adjustment(C<->D)}

\description{
 output file about predicting individual PK parameters of vancomycin by single conc. and dose adjustment(C<->D).
}

\keyword{misc}