\encoding{UTF-8}

\name{War.model}

\alias{War.model}

\title{Algorithm to estimate individual PK/PD parameters of warfarin}

\description{
 Algorithm to estimate individual PK/PD parameters of warfarin by Bayesian technique(OpenBUGS).
}

\references{
Dalere, GM, et al., A graphic for Warfarin Dosage Adjustment, Pharmacotherapy. 1999, 19(4):461-467 ; 
Theofanus TG, Barile RG. Multiple-dose kinetics of oral anticoagulants: methods of analysis and 
optimized dosing. J Pharm Sci 1973;62:261-6 ;
Bubp J, Coleman R, Mungall D, Raskob G, Hull R. Comparison of warfarin pharmacodynamic model predictions 
using Manchester comparative reagent vs Simplastin reagent. Clin Pharmacol Ther 1990;47:157.
}

\keyword{misc}