\encoding{UTF-8}

\name{War.more}

\alias{War.more}

\title{Function for individual dosage adjustment of warfarin}

\description{
 Function to predict individual INR of warfarin or to calculate their suggested doses 
 after perdicting individual PK parameters with single INR.
 }

\keyword{misc}