\encoding{UTF-8}

\name{warfarin}

\alias{Warfarin}

\title{Demo for the test runs to estimate individual PK/PD parameters of warfarin with one concentration}

\description{Demo for the test runs to estimate individual PK/PD parameters of warfarin with one concentration at steady-state including estimation of PK parameters}

\keyword{misc}