\name{Warss.output}

\alias{Warss.output}

\title{output file about predicting individual PK parameters of warfarin and dose adjustment(INR->D or D->INR)}

\description{
 output file about predicting individual PK parameters of warfarin by single INR and dose adjustment (INR->D or D->INR).
}

\keyword{misc}