\name{Warsscd.output}

\alias{Warsscd.output}

\title{output file about predicting individual PK parameters of warfarin and dose adjustment(INR<->D)}

\description{
 output file about predicting individual PK parameters of warfarin by single INR and dose adjustment (INR<->D).
}

\keyword{misc}