% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{zero_to_one}
\alias{zero_to_one}
\title{Zero to One Transform a Vector}
\usage{
zero_to_one(data)
}
\arguments{
\item{data}{-- A vector of gene expression values.}
}
\value{
the transformed vector.
}
\description{
A data vector is transformed so that all values are in range [0,1].
}

