% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{zero_to_one_transform}
\alias{zero_to_one_transform}
\title{Zero to One Transformation}
\usage{
zero_to_one_transform(datatable)
}
\arguments{
\item{data}{-- A data.table of values. The first column should be
gene symbols and the rest should be expression values.}
}
\value{
the transformed datatable.
}
\description{
The input is standardized to the range [0,1] by row.
}

