\name{Ami.mm}

\alias{Ami.mm}

\title{Algorithm to estimate many subjects' PK parameters of aminoglycoside with different concentrations}

\description{
 Algorithm to estimate many subjects' PK parameters of aminoglycoside with different 
 concentrations at steady-state.
}

\references{
Burton, M. E., M. S. Chow, et al. (1986). Accuracy of Bayesian and Sawchuk-Zaske dosing 
methods for gentamicin. Clin Pharm 5(2): 143-9.
}

\keyword{misc}