\name{Amianhinfusion.more}

\alias{Amianhinfusion.more}

\title{Function for individual dosage adjustment of iv infusion aminophylline anhydrous with loading dose }

\description{
 Function to predict individual concentrations of iv infusion aminophylline anhydrous 
 with loading dose or to calculate their suggested doses after perdicting individual 
 PK parameters with single concentration.
}

\keyword{misc}