\encoding{UTF-8}

\name{Amidihinfusion.more}

\alias{Amidihinfusion.more}

\title{Function for individual dosage adjustment of iv infusion aminophylline dihydrous with loading dose }

\description{
 Function to predict individual concentrations of iv infusion Aminophylline dihydrous 
 with loading dose or to calculate their suggested doses after perdicting individual 
 PK parameters with single concentration. 
}


\keyword{misc}