\encoding{UTF-8}

\name{PheDemo}

\alias{PheDemo}

\title{Function for the test runs to estimate individual PK parameters of phenytoin with one concentration}

\description{Function for the test runs to estimate individual PK parameters of phenytoin with one concentration at steady-state including estimation of PK parameters}

\keyword{misc}