\encoding{UTF-8}

\name{tdm}

\alias{tdm}

\title{Function of Therapeutic drug monitoring}

\description{
tdm is desgined to estimate individual PK parameters with one or more concentrations using
JAGS (Just Another Gibbs Sampling) interfaced through the package - rjags.
It also can calculate a suggested dose for a target drug conc. and vice versa.
}

\keyword{misc}
