% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tdm.R
\name{tdm}
\alias{tdm}
\title{TDM Transformation on a Single Value}
\usage{
tdm(value, iqr, old_min, old_max, old_third_q, old_first_q, ref_min, ref_max,
  ref_third_q, ref_first_q, inv_reference = TRUE, log_target = TRUE)
}
\arguments{
\item{value}{-- A single value from a distribution that is being
TDM transformed.}

\item{iqr}{-- inter-quartile-range of distribution being transformed}

\item{old_min}{-- min of distribution being transformed}

\item{old_max}{-- max of distribution being transformed}

\item{old_third_q}{-- third quartile of distribution being transformed}

\item{old_first_q}{-- first quartile of distribution being transformed}

\item{ref_min}{-- min of reference distribution}

\item{ref_max}{-- max of reference distribution}

\item{ref_third_q}{-- third quartile of reference distribution}

\item{ref_first_q}{-- first quartile of reference distribution}

\item{log_target}{-- logical indicating if target distribution should be logged}

\item{log_reference}{-- logical indcating if reference distribution was logged
before values were extracted}
}
\value{
TDM of the value passed.
}
\description{
Applies TDM transformation to a single value.
}

